backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
argon2-cffi==23.1.0
gunicorn==22.0.0
whitenoise==6.7.0
